requirements.txt
streamlit==1.28.0
pandas==2.1.0
python-dotenv==1.0.0
groq==0.30.0
plotly==5.17.0